move some methods back to configRules
